Per Provider Packaging
 - Added separated projects for SqlServer, MySql, PostgreSql and SQLite;
 - Moved all members from SqlAdapters folder for target provider project;
 - Reviewed all dependencies by EfCore provider;
 - Created IDbServer interface to be implemented by each EfCore provider, supplying instances for Adapter, Dialect, and some other features;
 - Added SqlAdaptersMapping.DbServer static property for setting up the desired provider to be used;
 - Refactored code that depends to call members from providers on core project.
