+++ EFCore.BulkExtensions.MySql/MySqlAdapter.cs
@@ -1,23 +1,21 @@
-﻿using EFCore.BulkExtensions.Helpers;
-using EFCore.BulkExtensions.SqlAdapters;
-using Microsoft.Data.SqlClient;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Storage;
-using NetTopologySuite.Geometries;
-using NetTopologySuite.IO;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Data;
+using System.Data.Common;
 using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Storage;
+using MySqlConnector;
+using NetTopologySuite.Geometries;
 
-namespace EFCore.BulkExtensions.SQLAdapters.SQLServer;
+namespace EFCore.BulkExtensions.SqlAdapters.MySql;
 
 /// <inheritdoc/>
-public class SqlOperationsServerAdapter: ISqlOperationsAdapter
+public class MySqlAdapter : ISqlOperationsAdapter
 {
     /// <inheritdoc/>
     #region Methods
@@ -28,14 +26,15 @@ public class SqlOperationsServerAdapter: ISqlOperationsAdapter
     }
 
     /// <inheritdoc/>
-    public async Task InsertAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, Action<decimal>? progress, CancellationToken cancellationToken)
+    public async Task InsertAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, Action<decimal>? progress,
+        CancellationToken cancellationToken)
     {
-        await InsertAsync(context, type, entities, tableInfo, progress, isAsync: true, cancellationToken).ConfigureAwait(false);
+        await InsertAsync(context, type, entities, tableInfo, progress, isAsync: false, CancellationToken.None).ConfigureAwait(false);
     }
-    // Public Async and NonAsync are merged into single operation flow with protected method using arg: bool isAsync (keeps code DRY)
-    // https://docs.microsoft.com/en-us/archive/msdn-magazine/2015/july/async-programming-brownfield-async-development#the-flag-argument-hack
+
     /// <inheritdoc/>
-    protected static async Task InsertAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, Action<decimal>? progress, bool isAsync, CancellationToken cancellationToken)
+    protected static async Task InsertAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo,
+        Action<decimal>? progress, bool isAsync, CancellationToken cancellationToken)
     {
         tableInfo.CheckToSetIdentityForPreserveOrder(tableInfo, entities);
         if (isAsync)
@@ -47,50 +46,27 @@ public class SqlOperationsServerAdapter: ISqlOperationsAdapter
             context.Database.OpenConnection();
         }
         var connection = context.GetUnderlyingConnection(tableInfo.BulkConfig);
-
         try
         {
             var transaction = context.Database.CurrentTransaction;
-
-            using var sqlBulkCopy = GetSqlBulkCopy((SqlConnection)connection, transaction, tableInfo.BulkConfig);
-            bool setColumnMapping = false;
-            tableInfo.SetSqlBulkCopyConfig(sqlBulkCopy, entities, setColumnMapping, progress);
+            var mySqlBulkCopy = GetMySqlBulkCopy((MySqlConnection)connection, transaction, tableInfo.BulkConfig);
             try
             {
-                var dataTable = GetDataTable(context, type, entities, sqlBulkCopy, tableInfo);
+                SetMySqlBulkCopyConfig(mySqlBulkCopy, tableInfo);
+
+                var dataTable = GetDataTable(context, type, entities, mySqlBulkCopy, tableInfo);
                 if (isAsync)
                 {
-                    await sqlBulkCopy.WriteToServerAsync(dataTable, cancellationToken).ConfigureAwait(false);
+                    await mySqlBulkCopy.WriteToServerAsync(dataTable, cancellationToken).ConfigureAwait(false);
                 }
                 else
                 {
-                    sqlBulkCopy.WriteToServer(dataTable);
+                    mySqlBulkCopy.WriteToServer(dataTable);
                 }
             }
-            catch (InvalidOperationException ex)
+            catch (InvalidOperationException e)
             {
-                if (ex.Message.Contains(BulkExceptionMessage.ColumnMappingNotMatch))
-                {
-                    bool tableExist = isAsync ? await TableInfo.CheckTableExistAsync(context, tableInfo, isAsync: true, cancellationToken).ConfigureAwait(false)
-                                                    : TableInfo.CheckTableExistAsync(context, tableInfo, isAsync: false, cancellationToken).GetAwaiter().GetResult();
-
-                    if (!tableExist)
-                    {
-                        var sqlCreateTableCopy = SqlQueryBuilder.CreateTableCopy(tableInfo.FullTableName, tableInfo.FullTempTableName, tableInfo);
-                        var sqlDropTable = SqlQueryBuilder.DropTable(tableInfo.FullTempTableName, tableInfo.BulkConfig.UseTempDB);
-
-                        if (isAsync)
-                        {
-                            await context.Database.ExecuteSqlRawAsync(sqlCreateTableCopy, cancellationToken).ConfigureAwait(false);
-                            await context.Database.ExecuteSqlRawAsync(sqlDropTable, cancellationToken).ConfigureAwait(false);
-                        }
-                        else
-                        {
-                            context.Database.ExecuteSqlRaw(sqlCreateTableCopy);
-                            context.Database.ExecuteSqlRaw(sqlDropTable);
-                        }
-                    }
-                }
+                Console.WriteLine(e);
                 throw;
             }
         }
@@ -110,44 +86,34 @@ public class SqlOperationsServerAdapter: ISqlOperationsAdapter
             tableInfo.CheckToSetIdentityForPreserveOrder(tableInfo, entities, reset: true);
         }
     }
-
     /// <inheritdoc/>
-    public void Merge<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, OperationType operationType, Action<decimal>? progress) where T : class
+    public void Merge<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, OperationType operationType,
+        Action<decimal>? progress) where T : class
     {
         MergeAsync(context, type, entities, tableInfo, operationType, progress, isAsync: false, CancellationToken.None).GetAwaiter().GetResult();
     }
 
     /// <inheritdoc/>
-    public async Task MergeAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, OperationType operationType, Action<decimal>? progress, CancellationToken cancellationToken) where T : class
+    public async Task MergeAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, OperationType operationType,
+        Action<decimal>? progress, CancellationToken cancellationToken) where T : class
     {
-        await MergeAsync(context, type, entities, tableInfo, operationType, progress, isAsync: true, cancellationToken).ConfigureAwait(false);
+        await  MergeAsync(context, type, entities, tableInfo, operationType, progress, isAsync: true, CancellationToken.None).ConfigureAwait(false);
     }
 
     /// <inheritdoc/>
-    protected async Task MergeAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, OperationType operationType, Action<decimal>? progress, bool isAsync, CancellationToken cancellationToken) where T : class
+    protected async Task MergeAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo,
+        OperationType operationType, Action<decimal>? progress, bool isAsync, CancellationToken cancellationToken)
+        where T : class
     {
-        var entityPropertyWithDefaultValue = entities.GetPropertiesWithDefaultValue(type, tableInfo);
-
+        //Because of using temp table in case of update, we need to access created temp table in Insert method.
+        var hasExistingTransaction = context.Database.CurrentTransaction != null;
+        var transaction = context.Database.CurrentTransaction ?? (isAsync ? await context.Database.BeginTransactionAsync(cancellationToken) : context.Database.BeginTransaction());
+        
         if (tableInfo.BulkConfig.CustomSourceTableName == null)
         {
             tableInfo.InsertToTempTable = true;
 
-            var dropTempTableIfExists = tableInfo.BulkConfig.UseTempDB;
-
-            if (dropTempTableIfExists)
-            {
-                var sqlDropTable = SqlQueryBuilder.DropTable(tableInfo.FullTempTableName, tableInfo.BulkConfig.UseTempDB);
-                if (isAsync)
-                {
-                    await context.Database.ExecuteSqlRawAsync(sqlDropTable, cancellationToken).ConfigureAwait(false);
-                }
-                else
-                {
-                    context.Database.ExecuteSqlRaw(sqlDropTable);
-                }
-            }
-
-            var sqlCreateTableCopy = SqlQueryBuilder.CreateTableCopy(tableInfo.FullTableName, tableInfo.FullTempTableName, tableInfo);
+            var sqlCreateTableCopy = SqlQueryBuilderMySql.CreateTableCopy(tableInfo.FullTableName, tableInfo.FullTempTableName, tableInfo.InsertToTempTable);
             if (isAsync)
             {
                 await context.Database.ExecuteSqlRawAsync(sqlCreateTableCopy, cancellationToken).ConfigureAwait(false);
@@ -156,101 +122,74 @@ public class SqlOperationsServerAdapter: ISqlOperationsAdapter
             {
                 context.Database.ExecuteSqlRaw(sqlCreateTableCopy);
             }
+        }
 
-            if (tableInfo.TimeStampColumnName != null)
+        bool doDropUniqueConstrain = false;
+        bool hasUniqueConstrain = false;
+        if (string.Join("_", tableInfo.EntityPKPropertyColumnNameDict.Keys.ToList()) == string.Join("_", tableInfo.PrimaryKeysPropertyColumnNameDict.Keys.ToList()))
+        {
+            hasUniqueConstrain = true; // ExplicitUniqueConstrain not required for PK
+        }
+        if (!hasUniqueConstrain)
+        {
+        // TODO 
+        //hasUniqueConstrain = await CheckHasExplicitUniqueConstrainAsync(context, connection, tableInfo, isAsync, cancellationToken);
+        // throws "The transaction associated with this command is not the connection’s active transaction";
+        // https://mysqlconnector.net/troubleshooting/transaction-usage/
+        }
+        if (!hasUniqueConstrain)
+        {
+            string createUniqueConstrain = SqlQueryBuilderMySql.CreateUniqueConstrain(tableInfo);
+            if (isAsync)
             {
-                var sqlAddColumn = SqlQueryBuilder.AddColumn(tableInfo.FullTempTableName, tableInfo.TimeStampColumnName, TableInfo.TimeStampOutColumnType);
-                if (isAsync)
-                {
-                    await context.Database.ExecuteSqlRawAsync(sqlAddColumn, cancellationToken).ConfigureAwait(false);
-                }
-                else
-                {
-                    context.Database.ExecuteSqlRaw(sqlAddColumn);
-                }
+                await context.Database.ExecuteSqlRawAsync(createUniqueConstrain, cancellationToken).ConfigureAwait(false);
             }
+            else
+            {
+                context.Database.ExecuteSqlRaw(createUniqueConstrain);
+            }
+            doDropUniqueConstrain = true;
         }
 
         if (tableInfo.CreatedOutputTable)
         {
-            var sqlCreateOutputTableCopy = SqlQueryBuilder.CreateTableCopy(tableInfo.FullTableName, tableInfo.FullTempOutputTableName, tableInfo, true);
+            tableInfo.InsertToTempTable = true;
+            var sqlCreateOutputTableCopy = SqlQueryBuilderMySql.CreateTableCopy(tableInfo.FullTableName,
+                tableInfo.FullTempOutputTableName, tableInfo.InsertToTempTable);
             if (isAsync)
             {
-                await context.Database.ExecuteSqlRawAsync(sqlCreateOutputTableCopy, cancellationToken).ConfigureAwait(false);
+                await context.Database.ExecuteSqlRawAsync(sqlCreateOutputTableCopy, cancellationToken)
+                    .ConfigureAwait(false);
             }
             else
             {
                 context.Database.ExecuteSqlRaw(sqlCreateOutputTableCopy);
             }
-
-            if (tableInfo.TimeStampColumnName != null)
-            {
-                var sqlAddColumn = SqlQueryBuilder.AddColumn(tableInfo.FullTempOutputTableName, tableInfo.TimeStampColumnName, TableInfo.TimeStampOutColumnType);
-                if (isAsync)
-                {
-                    await context.Database.ExecuteSqlRawAsync(sqlAddColumn, cancellationToken).ConfigureAwait(false);
-                }
-                else
-                {
-                    context.Database.ExecuteSqlRaw(sqlAddColumn);
-                }
-            }
-
-            if (operationType == OperationType.InsertOrUpdateOrDelete)
-            {
-                // Output returns all changes including Deleted rows with all NULL values, so if TempOutput.Id col not Nullable it breaks
-                var sqlAlterTableColumnsToNullable = SqlQueryBuilder.AlterTableColumnsToNullable(tableInfo.FullTempOutputTableName, tableInfo);
-                if (isAsync)
-                {
-                    await context.Database.ExecuteSqlRawAsync(sqlAlterTableColumnsToNullable, cancellationToken).ConfigureAwait(false);
-                }
-                else
-                {
-                    context.Database.ExecuteSqlRaw(sqlAlterTableColumnsToNullable);
-                }
-            }
         }
 
-        bool keepIdentity = tableInfo.BulkConfig.SqlBulkCopyOptions.HasFlag(SqlBulkCopyOptions.KeepIdentity);
-        try
+        if (tableInfo.BulkConfig.CustomSourceTableName == null)
         {
-            if (tableInfo.BulkConfig.CustomSourceTableName == null)
-            {
-                if (isAsync)
-                {
-                    await InsertAsync(context, type, entities, tableInfo, progress, cancellationToken).ConfigureAwait(false);
-                }
-                else
-                {
-                    Insert(context, type, entities, tableInfo, progress);
-                }
-            }
-
-            if (keepIdentity && tableInfo.HasIdentity)
-            {
-                var sqlSetIdentityInsertTrue = SqlQueryBuilder.SetIdentityInsert(tableInfo.FullTableName, true);
-                if (isAsync)
-                {
-                    await context.Database.OpenConnectionAsync(cancellationToken).ConfigureAwait(false);
-                    await context.Database.ExecuteSqlRawAsync(sqlSetIdentityInsertTrue, cancellationToken).ConfigureAwait(false);
-                }
-                else
-                {
-                    context.Database.OpenConnection();
-                    context.Database.ExecuteSqlRaw(sqlSetIdentityInsertTrue);
-                }
-            }
-
-            var (sql, parameters) = SqlQueryBuilder.MergeTable<T>(context, tableInfo, operationType, entityPropertyWithDefaultValue);
             if (isAsync)
             {
-                await context.Database.ExecuteSqlRawAsync(sql, parameters, cancellationToken).ConfigureAwait(false);
+                await InsertAsync(context, type, entities, tableInfo, progress, cancellationToken).ConfigureAwait(false);
             }
             else
             {
-                context.Database.ExecuteSqlRaw(sql, parameters);
+                Insert(context, type, entities, tableInfo, progress);
             }
+        }
 
+        try
+        {
+            var sqlMergeTable = SqlQueryBuilderMySql.MergeTable<T>(tableInfo, operationType);
+            if (isAsync)
+            {
+                await context.Database.ExecuteSqlRawAsync(sqlMergeTable, cancellationToken).ConfigureAwait(false);
+            }
+            else
+            {
+                context.Database.ExecuteSqlRaw(sqlMergeTable);
+            }
             if (tableInfo.CreatedOutputTable)
             {
                 if (isAsync)
@@ -262,14 +201,39 @@ public class SqlOperationsServerAdapter: ISqlOperationsAdapter
                     tableInfo.LoadOutputDataAsync(context, type, entities, tableInfo, isAsync: false, cancellationToken).GetAwaiter().GetResult();
                 }
             }
+            if (hasExistingTransaction == false && !tableInfo.BulkConfig.IncludeGraph)
+            {
+                if (isAsync)
+                {
+                    await transaction.CommitAsync(cancellationToken);
+                }
+                else
+                {
+                    transaction.Commit();
+                }
+            }
         }
         finally
         {
+            if (doDropUniqueConstrain)
+            {
+                string dropUniqueConstrain = SqlQueryBuilderMySql.DropUniqueConstrain(tableInfo);
+                if (isAsync)
+                {
+                    await context.Database.ExecuteSqlRawAsync(dropUniqueConstrain, cancellationToken)
+                        .ConfigureAwait(false);
+                }
+                else
+                {
+                    context.Database.ExecuteSqlRaw(dropUniqueConstrain);
+                }
+            }
+
             if (!tableInfo.BulkConfig.UseTempDB)
             {
                 if (tableInfo.CreatedOutputTable)
                 {
-                    var sqlDropOutputTable = SqlQueryBuilder.DropTable(tableInfo.FullTempOutputTableName, tableInfo.BulkConfig.UseTempDB);
+                    var sqlDropOutputTable = SqlQueryBuilderMySql.DropTable(tableInfo.FullTempOutputTableName, tableInfo.InsertToTempTable);
                     if (isAsync)
                     {
                         await context.Database.ExecuteSqlRawAsync(sqlDropOutputTable, cancellationToken).ConfigureAwait(false);
@@ -282,7 +246,7 @@ public class SqlOperationsServerAdapter: ISqlOperationsAdapter
                 }
                 if (tableInfo.BulkConfig.CustomSourceTableName == null)
                 {
-                    var sqlDropTable = SqlQueryBuilder.DropTable(tableInfo.FullTempTableName, tableInfo.BulkConfig.UseTempDB);
+                    var sqlDropTable = SqlQueryBuilderMySql.DropTable(tableInfo.FullTempTableName, tableInfo.InsertToTempTable);
                     if (isAsync)
                     {
                         await context.Database.ExecuteSqlRawAsync(sqlDropTable, cancellationToken).ConfigureAwait(false);
@@ -292,145 +256,115 @@ public class SqlOperationsServerAdapter: ISqlOperationsAdapter
                         context.Database.ExecuteSqlRaw(sqlDropTable);
                     }
                 }
-            }
-
-            if (keepIdentity && tableInfo.HasIdentity)
-            {
-                var sqlSetIdentityInsertFalse = SqlQueryBuilder.SetIdentityInsert(tableInfo.FullTableName, false);
-                if (isAsync)
+                if (hasExistingTransaction == false && !tableInfo.BulkConfig.IncludeGraph)
                 {
-                    await context.Database.ExecuteSqlRawAsync(sqlSetIdentityInsertFalse, cancellationToken).ConfigureAwait(false);
+                    if (isAsync)
+                    {
+                        await transaction.DisposeAsync();
+                    }
+                    else
+                    {
+                        transaction.Dispose();
+                    }
                 }
-                else
-                {
-                    context.Database.ExecuteSqlRaw(sqlSetIdentityInsertFalse);
-                }
-                context.Database.CloseConnection();
-            }
+            }    
         }
+        
+        
     }
-
     /// <inheritdoc/>
     public void Read<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, Action<decimal>? progress) where T : class
     {
-        ReadAsync(context, type, entities, tableInfo, progress, isAsync: false, CancellationToken.None).GetAwaiter().GetResult();
+        throw new NotImplementedException();
     }
 
     /// <inheritdoc/>
-    public async Task ReadAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, Action<decimal>? progress, CancellationToken cancellationToken) where T : class
+    public Task ReadAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, Action<decimal>? progress,
+        CancellationToken cancellationToken) where T : class
     {
-        await ReadAsync(context, type, entities, tableInfo, progress, isAsync: true, cancellationToken).ConfigureAwait(false);
-    }
-
-    /// <inheritdoc/>
-    protected async Task ReadAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, Action<decimal>? progress, bool isAsync, CancellationToken cancellationToken) where T : class
-    {
-        Dictionary<string, string> previousPropertyColumnNamesDict = tableInfo.ConfigureBulkReadTableInfo();
-
-        var sqlCreateTableCopy = SqlQueryBuilder.CreateTableCopy(tableInfo.FullTableName, tableInfo.FullTempTableName, tableInfo);
-        if (isAsync)
-        {
-            await context.Database.ExecuteSqlRawAsync(sqlCreateTableCopy, cancellationToken).ConfigureAwait(false);
-        }
-        else
-        {
-            context.Database.ExecuteSqlRaw(sqlCreateTableCopy);
-        }
-
-        try
-        {
-            if (isAsync)
-            {
-                await InsertAsync(context, type, entities, tableInfo, progress, cancellationToken).ConfigureAwait(false);
-            }
-            else
-            {
-                InsertAsync(context, type, entities, tableInfo, progress, isAsync: false, cancellationToken).GetAwaiter().GetResult();
-            }
-
-            tableInfo.PropertyColumnNamesDict = tableInfo.OutputPropertyColumnNamesDict;
-
-            var sqlSelectJoinTable = SqlQueryBuilder.SelectJoinTable(tableInfo);
-
-            tableInfo.PropertyColumnNamesDict = previousPropertyColumnNamesDict; // TODO Consider refactor and integrate with TimeStampPropertyName, also check for Calculated props.
-                                                                                 // Output only PropertisToInclude and for getting Id with SetOutputIdentity
-            if (tableInfo.TimeStampPropertyName != null && !tableInfo.PropertyColumnNamesDict.ContainsKey(tableInfo.TimeStampPropertyName) && tableInfo.TimeStampColumnName is not null)
-            {
-                tableInfo.PropertyColumnNamesDict.Add(tableInfo.TimeStampPropertyName, tableInfo.TimeStampColumnName);
-            }
-
-            List<T> existingEntities = tableInfo.LoadOutputEntities<T>(context, type, sqlSelectJoinTable);
-            tableInfo.UpdateReadEntities(entities, existingEntities, context);
-
-            if (tableInfo.TimeStampPropertyName != null && !tableInfo.PropertyColumnNamesDict.ContainsKey(tableInfo.TimeStampPropertyName))
-            {
-                tableInfo.PropertyColumnNamesDict.Remove(tableInfo.TimeStampPropertyName);
-            }
-        }
-        finally
-        {
-            if (!tableInfo.BulkConfig.UseTempDB)
-            {
-                var sqlDropTable = SqlQueryBuilder.DropTable(tableInfo.FullTempTableName, tableInfo.BulkConfig.UseTempDB);
-                if (isAsync)
-                {
-                    await context.Database.ExecuteSqlRawAsync(sqlDropTable, cancellationToken).ConfigureAwait(false);
-                }
-                else
-                {
-                    context.Database.ExecuteSqlRaw(sqlDropTable);
-                }
-            }
-        }
+        throw new NotImplementedException();
     }
 
     /// <inheritdoc/>
     public void Truncate(DbContext context, TableInfo tableInfo)
     {
-        var sqlTruncateTable = SqlQueryBuilder.TruncateTable(tableInfo.FullTableName);
-        context.Database.ExecuteSqlRaw(sqlTruncateTable);
+        throw new NotImplementedException();
     }
 
     /// <inheritdoc/>
-    public async Task TruncateAsync(DbContext context, TableInfo tableInfo, CancellationToken cancellationToken)
+    public Task TruncateAsync(DbContext context, TableInfo tableInfo, CancellationToken cancellationToken)
     {
-        var sqlTruncateTable = SqlQueryBuilder.TruncateTable(tableInfo.FullTableName);
-        await context.Database.ExecuteSqlRawAsync(sqlTruncateTable, cancellationToken).ConfigureAwait(false);
+        throw new NotImplementedException();
     }
     #endregion
-
     #region Connection
-    private static SqlBulkCopy GetSqlBulkCopy(SqlConnection sqlConnection, IDbContextTransaction? transaction, BulkConfig config)
+
+    internal static async Task<(MySqlConnection, bool)> OpenAndGetMySqlConnectionAsync(DbContext context, CancellationToken cancellationToken)
     {
-        var sqlTransaction = transaction == null ? null : (SqlTransaction)transaction.GetUnderlyingTransaction(config);
-        var sqlBulkCopy = new SqlBulkCopy(sqlConnection, config.SqlBulkCopyOptions, sqlTransaction);
-        if (config.SqlBulkCopyColumnOrderHints != null)
+        bool closeConnectionInternally = false;
+        var mySqlConnection = (MySqlConnection)context.Database.GetDbConnection();
+        if (mySqlConnection.State != ConnectionState.Open)
         {
-            foreach(var hint in config.SqlBulkCopyColumnOrderHints)
-                sqlBulkCopy.ColumnOrderHints.Add(hint);
+            await mySqlConnection.OpenAsync(cancellationToken).ConfigureAwait(false);
+            closeConnectionInternally = true;
         }
-        return sqlBulkCopy;
+        return (mySqlConnection, closeConnectionInternally);
+    }
+
+    internal static (MySqlConnection, bool) OpenAndGetMySqlConnection(DbContext context)
+    {
+        bool closeConnectionInternally = false;
+        var myConnection = (MySqlConnection)context.Database.GetDbConnection();
+        if (myConnection.State != ConnectionState.Open)
+        {
+            myConnection.Open();
+            closeConnectionInternally = true;
+        }
+        return (myConnection, closeConnectionInternally);
     }
     #endregion
-    
+    #region Connection
+    private static MySqlBulkCopy GetMySqlBulkCopy(MySqlConnection mySqlConnection, IDbContextTransaction? transaction, BulkConfig config)
+    {
+        var mySqlTransaction = transaction == null ? null : (MySqlTransaction)transaction.GetUnderlyingTransaction(config);
+        var mySqlBulkCopy = new MySqlBulkCopy(mySqlConnection, mySqlTransaction);
+        
+        return mySqlBulkCopy;
+    }
+    /// <summary>
+    /// Supports <see cref="MySqlConnector.MySqlBulkCopy"/>
+    /// </summary>
+    /// <param name="mySqlBulkCopy"></param>
+    /// <param name="tableInfo"></param>
+    private static void SetMySqlBulkCopyConfig(MySqlBulkCopy mySqlBulkCopy, TableInfo tableInfo)
+    {
+        mySqlBulkCopy.DestinationTableName = tableInfo.InsertToTempTable ? tableInfo.FullTempTableName : tableInfo.FullTableName;
+        mySqlBulkCopy.DestinationTableName = mySqlBulkCopy.DestinationTableName.Replace("[", "").Replace("]", "");
+        mySqlBulkCopy.NotifyAfter = tableInfo.BulkConfig.NotifyAfter ?? tableInfo.BulkConfig.BatchSize;
+        mySqlBulkCopy.BulkCopyTimeout = tableInfo.BulkConfig.BulkCopyTimeout ?? mySqlBulkCopy.BulkCopyTimeout;
+    }
+
+    #endregion
     #region DataTable
     /// <summary>
-    /// Supports <see cref="SqlBulkCopy"/>
+    /// Supports <see cref="MySqlBulkCopy"/>
     /// </summary>
     /// <typeparam name="T"></typeparam>
     /// <param name="context"></param>
     /// <param name="type"></param>
     /// <param name="entities"></param>
-    /// <param name="sqlBulkCopy"></param>
+    /// <param name="mySqlBulkCopy"></param>
     /// <param name="tableInfo"></param>
     /// <returns></returns>
-    public static DataTable GetDataTable<T>(DbContext context, Type type, IList<T> entities, SqlBulkCopy sqlBulkCopy, TableInfo tableInfo)
+    public static DataTable GetDataTable<T>(DbContext context, Type type, IList<T> entities, MySqlBulkCopy mySqlBulkCopy, TableInfo tableInfo)
     {
         DataTable dataTable = InnerGetDataTable(context, ref type, entities, tableInfo);
 
+        int sourceOrdinal = 0;
         foreach (DataColumn item in dataTable.Columns)  //Add mapping
         {
-            sqlBulkCopy.ColumnMappings.Add(item.ColumnName, item.ColumnName);
+            mySqlBulkCopy.ColumnMappings.Add(new MySqlBulkCopyColumnMapping(sourceOrdinal,item.ColumnName));
+            sourceOrdinal++;
         }
         return dataTable;
     }
@@ -450,10 +384,9 @@ public class SqlOperationsServerAdapter: ISqlOperationsAdapter
         var columnsDict = new Dictionary<string, object?>();
         var ownedEntitiesMappedProperties = new HashSet<string>();
 
-        var databaseType = SqlAdaptersMapping.GetDatabaseType(context);
-        var isSqlServer = databaseType == DbServer.SQLServer;
-        var sqlServerBytesWriter = new SqlServerBytesWriter();
-
+        var databaseType = SqlAdaptersMapping.GetDatabaseType();
+        var isMySql = databaseType == DbServer.MySQL;
+        
         var objectIdentifier = tableInfo.ObjectIdentifier;
         type = tableInfo.HasAbstractList ? entities[0]!.GetType() : type;
         var entityType = context.Model.FindEntityType(type) ?? throw new ArgumentException($"Unable to determine entity type from given type - {type.Name}");
@@ -495,7 +428,7 @@ public class SqlOperationsServerAdapter: ISqlOperationsAdapter
                     propertyType = underlyingType;
                 }
 
-                if (isSqlServer && (propertyType == typeof(Geometry) || propertyType.IsSubclassOf(typeof(Geometry))))
+                if (isMySql && (propertyType == typeof(Geometry) || propertyType.IsSubclassOf(typeof(Geometry))))
                 {
                     propertyType = typeof(byte[]);
                     tableInfo.HasSpatialType = true;
@@ -504,7 +437,7 @@ public class SqlOperationsServerAdapter: ISqlOperationsAdapter
                         throw new InvalidOperationException("OnCompare properties Config can not be set for Entity with Spatial types like 'Geometry'");
                     }
                 }
-                if (isSqlServer && (propertyType == typeof(HierarchyId) || propertyType.IsSubclassOf(typeof(HierarchyId))))
+                if (isMySql && (propertyType == typeof(HierarchyId) || propertyType.IsSubclassOf(typeof(HierarchyId))))
                 {
                     propertyType = typeof(byte[]);
                 }
@@ -534,12 +467,12 @@ public class SqlOperationsServerAdapter: ISqlOperationsAdapter
                     propertyType = underlyingType;
                 }
 
-                if (propertyType == typeof(Geometry) && isSqlServer)
+                if (propertyType == typeof(Geometry) && isMySql)
                 {
                     propertyType = typeof(byte[]);
                 }
 
-                if (propertyType == typeof(HierarchyId) && isSqlServer)
+                if (propertyType == typeof(HierarchyId) && isMySql)
                 {
                     propertyType = typeof(byte[]);
                 }
@@ -590,31 +523,14 @@ public class SqlOperationsServerAdapter: ISqlOperationsAdapter
                             {
                                 var convertor = tableInfo.ConvertibleColumnConverterDict[propertyName];
                                 var underlyingType = Nullable.GetUnderlyingType(convertor.ProviderClrType) ?? convertor.ProviderClrType;
-          
                                 dataTable.Columns.Add(columnName, underlyingType);
                             }
                             else
                             {
                                 var ownedPropertyType = Nullable.GetUnderlyingType(innerProperty.PropertyType) ?? innerProperty.PropertyType;
-                                
-                                if (isSqlServer && (ownedPropertyType == typeof(Geometry) || ownedPropertyType.IsSubclassOf(typeof(Geometry))))
-                                {
-                                    ownedPropertyType = typeof(byte[]);
-                                    tableInfo.HasSpatialType = true;
-                                    if (tableInfo.BulkConfig.PropertiesToIncludeOnCompare != null || tableInfo.BulkConfig.PropertiesToIncludeOnCompare != null)
-                                    {
-                                        throw new InvalidOperationException("OnCompare properties Config can not be set for Entity with Spatial types like 'Geometry'");
-                                    }
-                                }
-
-                                if (isSqlServer && (ownedPropertyType == typeof(HierarchyId) || ownedPropertyType.IsSubclassOf(typeof(HierarchyId))))
-                                {
-                                    ownedPropertyType = typeof(byte[]);
-                                }
-
                                 dataTable.Columns.Add(columnName, ownedPropertyType);
                             }
-  ;
+
                             columnsDict.Add(property.Name + "_" + innerProperty.Name, null);
                         }
                     }
@@ -644,12 +560,12 @@ public class SqlOperationsServerAdapter: ISqlOperationsAdapter
                     propertyType = underlyingType;
                 }
 
-                if (isSqlServer && (propertyType == typeof(Geometry) || propertyType.IsSubclassOf(typeof(Geometry))))
+                if (isMySql && (propertyType == typeof(Geometry) || propertyType.IsSubclassOf(typeof(Geometry))))
                 {
                     propertyType = typeof(byte[]);
                 }
 
-                if (isSqlServer && (propertyType == typeof(HierarchyId) || propertyType.IsSubclassOf(typeof(HierarchyId))))
+                if (isMySql && (propertyType == typeof(HierarchyId) || propertyType.IsSubclassOf(typeof(HierarchyId))))
                 {
                     propertyType = typeof(byte[]);
                 }
@@ -686,7 +602,7 @@ public class SqlOperationsServerAdapter: ISqlOperationsAdapter
                     && tableInfo.DefaultValueProperties.Contains(property.Name);
 
                 if (tableInfo.BulkConfig.DateTime2PrecisionForceRound
-                    && isSqlServer
+                    && isMySql
                     && tableInfo.DateTime2PropertiesPrecisionLessThen7Dict.ContainsKey(property.Name))
                 {
                     DateTime? dateTimePropertyValue = (DateTime?)propertyValue;
@@ -711,20 +627,9 @@ public class SqlOperationsServerAdapter: ISqlOperationsAdapter
                     string columnName = tableInfo.ConvertiblePropertyColumnDict[property.Name];
                     propertyValue = tableInfo.ConvertibleColumnConverterDict[columnName].ConvertToProvider.Invoke(propertyValue);
                 }
+                //TODO: Hamdling special types
 
-                if (tableInfo.HasSpatialType && propertyValue is Geometry geometryValue)
-                {
-                    geometryValue.SRID = tableInfo.BulkConfig.SRID;
-
-                    if (tableInfo.PropertyColumnNamesDict.ContainsKey(property.Name))
-                    {
-                        sqlServerBytesWriter.IsGeography = tableInfo.ColumnNamesTypesDict[tableInfo.PropertyColumnNamesDict[property.Name]] == "geography"; // "geography" type is default, otherwise it's "geometry" type
-                    }
-
-                    propertyValue = sqlServerBytesWriter.Write(geometryValue);
-                }
-
-                if (propertyValue is HierarchyId hierarchyValue && isSqlServer)
+                if (propertyValue is HierarchyId hierarchyValue && isMySql)
                 {
                     using MemoryStream memStream = new();
                     using BinaryWriter binWriter = new(memStream);
@@ -762,17 +667,6 @@ public class SqlOperationsServerAdapter: ISqlOperationsAdapter
                         {
                             var converter = tableInfo.ConvertibleColumnConverterDict[columnName];
                             columnsDict[columnName] = ownedPropertyValue == null ? null : converter.ConvertToProvider.Invoke(ownedPropertyValue);
-                        }
-                        else if(tableInfo.HasSpatialType && ownedPropertyValue is Geometry ownedGeometryValue)
-                        {
-                            ownedGeometryValue.SRID = tableInfo.BulkConfig.SRID;
-
-                            if (tableInfo.PropertyColumnNamesDict.ContainsKey(property.Name))
-                            {
-                                sqlServerBytesWriter.IsGeography = tableInfo.ColumnNamesTypesDict[tableInfo.PropertyColumnNamesDict[property.Name]] == "geography"; // "geography" type is default, otherwise it's "geometry" type
-                            }
-
-                            columnsDict[columnName] = sqlServerBytesWriter.Write(ownedGeometryValue);
                         }
                         else
                         {
