+++ EFCore.BulkExtensions.MySql/MySqlDialect.cs
@@ -1,0 +1,9 @@
+﻿using EFCore.BulkExtensions.SqlAdapters;
+
+namespace EFCore.BulkExtensions.SqlAdapters.MySql;
+
+/// <inheritdoc/>
+public class MySqlDialect : SqlDefaultDialect
+{
+    
+}
