+++ EFCore.BulkExtensions.PostgreSql/EFCore.BulkExtensions.PostgreSql.csproj
@@ -1,0 +1,44 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+      <TargetFramework>net6.0</TargetFramework>
+      <Title>EFCore.BulkExtensions.PostgreSql</Title>
+      <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+      <Version>6.5.6</Version>
+      <Nullable>enable</Nullable>
+      <Authors>borisdj</Authors>
+      <Description>EntityFramework EF Core Bulk Batch Extensions for Insert Update Delete Read (CRUD) operations on SQL Server, PostgreSQL, MySQL, SQLite</Description>
+      <PackageProjectUrl>https://github.com/borisdj/EFCore.BulkExtensions</PackageProjectUrl>
+      <Company />
+      <PackageIcon>EFCoreBulk.png</PackageIcon>
+      <PackageLicenseExpression>MIT</PackageLicenseExpression>
+      <PackageTags>EntityFrameworkCore Entity Framework Core EFCore EF Core SQL Server MySQL PostgreSQL SQLite Bulk Batch Extensions Insert Update Delete Read Truncate SaveChanges </PackageTags>
+      <PackageReleaseNotes>MySQL CRUD support, PG limit fix</PackageReleaseNotes>
+      <RepositoryUrl>https://github.com/borisdj/EFCore.BulkExtensions</RepositoryUrl>
+      <RepositoryType>Git</RepositoryType>
+      <AllowedOutputExtensionsInPackageBuildOutputFolder>$(AllowedOutputExtensionsInPackageBuildOutputFolder);.pdb</AllowedOutputExtensionsInPackageBuildOutputFolder>
+      <AssemblyVersion>6.5.6.0</AssemblyVersion>
+      <FileVersion>6.5.6.0</FileVersion>
+      <SignAssembly>true</SignAssembly>
+      <AssemblyOriginatorKeyFile>..\Keys\EFCore.BulkExtensions.snk</AssemblyOriginatorKeyFile>
+      <DelaySign>false</DelaySign>
+      <GenerateDocumentationFile>true</GenerateDocumentationFile>
+      <LangVersion>latest</LangVersion>
+      <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="6.0.6" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\EFCore.BulkExtensions\EFCore.BulkExtensions.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+      <None Include="..\EFCoreBulk.png">
+          <Pack>True</Pack>
+          <PackagePath></PackagePath>
+      </None>
+  </ItemGroup>
+</Project>
