+++ EFCore.BulkExtensions.PostgreSql/PostgreSqlDialect.cs
@@ -1,0 +1,6 @@
+﻿namespace EFCore.BulkExtensions.SqlAdapters.PostgreSql;
+
+/// <inheritdoc/>
+public class PostgreSqlDialect : SqlDefaultDialect
+{
+}
