+++ EFCore.BulkExtensions.SqlServer/SqlServerAdapter.cs
@@ -1,21 +1,23 @@
-﻿using System;
+﻿using EFCore.BulkExtensions.Helpers;
+using EFCore.BulkExtensions.SqlAdapters;
+using Microsoft.Data.SqlClient;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Storage;
+using NetTopologySuite.Geometries;
+using NetTopologySuite.IO;
+using System;
 using System.Collections.Generic;
 using System.Data;
-using System.Data.Common;
 using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Storage;
-using MySqlConnector;
-using NetTopologySuite.Geometries;
 
-namespace EFCore.BulkExtensions.SqlAdapters.MySql;
+namespace EFCore.BulkExtensions.SqlAdapters.SQLServer;
 
 /// <inheritdoc/>
-public class MySqLAdapter : ISqlOperationsAdapter
+public class SqlOperationsServerAdapter: ISqlOperationsAdapter
 {
     /// <inheritdoc/>
     #region Methods
@@ -26,15 +28,14 @@ public class MySqLAdapter : ISqlOperationsAdapter
     }
 
     /// <inheritdoc/>
-    public async Task InsertAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, Action<decimal>? progress,
-        CancellationToken cancellationToken)
+    public async Task InsertAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, Action<decimal>? progress, CancellationToken cancellationToken)
     {
-        await InsertAsync(context, type, entities, tableInfo, progress, isAsync: false, CancellationToken.None).ConfigureAwait(false);
+        await InsertAsync(context, type, entities, tableInfo, progress, isAsync: true, cancellationToken).ConfigureAwait(false);
     }
-
+    // Public Async and NonAsync are merged into single operation flow with protected method using arg: bool isAsync (keeps code DRY)
+    // https://docs.microsoft.com/en-us/archive/msdn-magazine/2015/july/async-programming-brownfield-async-development#the-flag-argument-hack
     /// <inheritdoc/>
-    protected static async Task InsertAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo,
-        Action<decimal>? progress, bool isAsync, CancellationToken cancellationToken)
+    protected static async Task InsertAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, Action<decimal>? progress, bool isAsync, CancellationToken cancellationToken)
     {
         tableInfo.CheckToSetIdentityForPreserveOrder(tableInfo, entities);
         if (isAsync)
@@ -46,27 +47,50 @@ public class MySqLAdapter : ISqlOperationsAdapter
             context.Database.OpenConnection();
         }
         var connection = context.GetUnderlyingConnection(tableInfo.BulkConfig);
+
         try
         {
             var transaction = context.Database.CurrentTransaction;
-            var mySqlBulkCopy = GetMySqlBulkCopy((MySqlConnection)connection, transaction, tableInfo.BulkConfig);
+
+            using var sqlBulkCopy = GetSqlBulkCopy((SqlConnection)connection, transaction, tableInfo.BulkConfig);
+            bool setColumnMapping = false;
+            SetSqlBulkCopyConfig(sqlBulkCopy, tableInfo, entities, setColumnMapping, progress);
             try
             {
-                tableInfo.SetMySqlBulkCopyConfig(mySqlBulkCopy);
-
-                var dataTable = GetDataTable(context, type, entities, mySqlBulkCopy, tableInfo);
+                var dataTable = GetDataTable(context, type, entities, sqlBulkCopy, tableInfo);
                 if (isAsync)
                 {
-                    await mySqlBulkCopy.WriteToServerAsync(dataTable, cancellationToken).ConfigureAwait(false);
+                    await sqlBulkCopy.WriteToServerAsync(dataTable, cancellationToken).ConfigureAwait(false);
                 }
                 else
                 {
-                    mySqlBulkCopy.WriteToServer(dataTable);
+                    sqlBulkCopy.WriteToServer(dataTable);
                 }
             }
-            catch (InvalidOperationException e)
+            catch (InvalidOperationException ex)
             {
-                Console.WriteLine(e);
+                if (ex.Message.Contains(BulkExceptionMessage.ColumnMappingNotMatch))
+                {
+                    bool tableExist = isAsync ? await TableInfo.CheckTableExistAsync(context, tableInfo, isAsync: true, cancellationToken).ConfigureAwait(false)
+                                                    : TableInfo.CheckTableExistAsync(context, tableInfo, isAsync: false, cancellationToken).GetAwaiter().GetResult();
+
+                    if (!tableExist)
+                    {
+                        var sqlCreateTableCopy = SqlQueryBuilder.CreateTableCopy(tableInfo.FullTableName, tableInfo.FullTempTableName, tableInfo);
+                        var sqlDropTable = SqlQueryBuilder.DropTable(tableInfo.FullTempTableName, tableInfo.BulkConfig.UseTempDB);
+
+                        if (isAsync)
+                        {
+                            await context.Database.ExecuteSqlRawAsync(sqlCreateTableCopy, cancellationToken).ConfigureAwait(false);
+                            await context.Database.ExecuteSqlRawAsync(sqlDropTable, cancellationToken).ConfigureAwait(false);
+                        }
+                        else
+                        {
+                            context.Database.ExecuteSqlRaw(sqlCreateTableCopy);
+                            context.Database.ExecuteSqlRaw(sqlDropTable);
+                        }
+                    }
+                }
                 throw;
             }
         }
@@ -86,34 +110,44 @@ public class MySqLAdapter : ISqlOperationsAdapter
             tableInfo.CheckToSetIdentityForPreserveOrder(tableInfo, entities, reset: true);
         }
     }
+
     /// <inheritdoc/>
-    public void Merge<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, OperationType operationType,
-        Action<decimal>? progress) where T : class
+    public void Merge<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, OperationType operationType, Action<decimal>? progress) where T : class
     {
         MergeAsync(context, type, entities, tableInfo, operationType, progress, isAsync: false, CancellationToken.None).GetAwaiter().GetResult();
     }
 
     /// <inheritdoc/>
-    public async Task MergeAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, OperationType operationType,
-        Action<decimal>? progress, CancellationToken cancellationToken) where T : class
+    public async Task MergeAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, OperationType operationType, Action<decimal>? progress, CancellationToken cancellationToken) where T : class
     {
-        await  MergeAsync(context, type, entities, tableInfo, operationType, progress, isAsync: true, CancellationToken.None).ConfigureAwait(false);
+        await MergeAsync(context, type, entities, tableInfo, operationType, progress, isAsync: true, cancellationToken).ConfigureAwait(false);
     }
 
     /// <inheritdoc/>
-    protected async Task MergeAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo,
-        OperationType operationType, Action<decimal>? progress, bool isAsync, CancellationToken cancellationToken)
-        where T : class
+    protected async Task MergeAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, OperationType operationType, Action<decimal>? progress, bool isAsync, CancellationToken cancellationToken) where T : class
     {
-        //Because of using temp table in case of update, we need to access created temp table in Insert method.
-        var hasExistingTransaction = context.Database.CurrentTransaction != null;
-        var transaction = context.Database.CurrentTransaction ?? (isAsync ? await context.Database.BeginTransactionAsync(cancellationToken) : context.Database.BeginTransaction());
-        
+        var entityPropertyWithDefaultValue = entities.GetPropertiesWithDefaultValue(type, tableInfo);
+
         if (tableInfo.BulkConfig.CustomSourceTableName == null)
         {
             tableInfo.InsertToTempTable = true;
 
-            var sqlCreateTableCopy = SqlQueryBuilderMySql.CreateTableCopy(tableInfo.FullTableName, tableInfo.FullTempTableName, tableInfo.InsertToTempTable);
+            var dropTempTableIfExists = tableInfo.BulkConfig.UseTempDB;
+
+            if (dropTempTableIfExists)
+            {
+                var sqlDropTable = SqlQueryBuilder.DropTable(tableInfo.FullTempTableName, tableInfo.BulkConfig.UseTempDB);
+                if (isAsync)
+                {
+                    await context.Database.ExecuteSqlRawAsync(sqlDropTable, cancellationToken).ConfigureAwait(false);
+                }
+                else
+                {
+                    context.Database.ExecuteSqlRaw(sqlDropTable);
+                }
+            }
+
+            var sqlCreateTableCopy = SqlQueryBuilder.CreateTableCopy(tableInfo.FullTableName, tableInfo.FullTempTableName, tableInfo);
             if (isAsync)
             {
                 await context.Database.ExecuteSqlRawAsync(sqlCreateTableCopy, cancellationToken).ConfigureAwait(false);
@@ -122,74 +156,101 @@ public class MySqLAdapter : ISqlOperationsAdapter
             {
                 context.Database.ExecuteSqlRaw(sqlCreateTableCopy);
             }
-        }
 
-        bool doDropUniqueConstrain = false;
-        bool hasUniqueConstrain = false;
-        if (string.Join("_", tableInfo.EntityPKPropertyColumnNameDict.Keys.ToList()) == string.Join("_", tableInfo.PrimaryKeysPropertyColumnNameDict.Keys.ToList()))
-        {
-            hasUniqueConstrain = true; // ExplicitUniqueConstrain not required for PK
-        }
-        if (!hasUniqueConstrain)
-        {
-        // TODO 
-        //hasUniqueConstrain = await CheckHasExplicitUniqueConstrainAsync(context, connection, tableInfo, isAsync, cancellationToken);
-        // throws "The transaction associated with this command is not the connection’s active transaction";
-        // https://mysqlconnector.net/troubleshooting/transaction-usage/
-        }
-        if (!hasUniqueConstrain)
-        {
-            string createUniqueConstrain = SqlQueryBuilderMySql.CreateUniqueConstrain(tableInfo);
-            if (isAsync)
+            if (tableInfo.TimeStampColumnName != null)
             {
-                await context.Database.ExecuteSqlRawAsync(createUniqueConstrain, cancellationToken).ConfigureAwait(false);
+                var sqlAddColumn = SqlQueryBuilder.AddColumn(tableInfo.FullTempTableName, tableInfo.TimeStampColumnName, TableInfo.TimeStampOutColumnType);
+                if (isAsync)
+                {
+                    await context.Database.ExecuteSqlRawAsync(sqlAddColumn, cancellationToken).ConfigureAwait(false);
+                }
+                else
+                {
+                    context.Database.ExecuteSqlRaw(sqlAddColumn);
+                }
             }
-            else
-            {
-                context.Database.ExecuteSqlRaw(createUniqueConstrain);
-            }
-            doDropUniqueConstrain = true;
         }
 
         if (tableInfo.CreatedOutputTable)
         {
-            tableInfo.InsertToTempTable = true;
-            var sqlCreateOutputTableCopy = SqlQueryBuilderMySql.CreateTableCopy(tableInfo.FullTableName,
-                tableInfo.FullTempOutputTableName, tableInfo.InsertToTempTable);
+            var sqlCreateOutputTableCopy = SqlQueryBuilder.CreateTableCopy(tableInfo.FullTableName, tableInfo.FullTempOutputTableName, tableInfo, true);
             if (isAsync)
             {
-                await context.Database.ExecuteSqlRawAsync(sqlCreateOutputTableCopy, cancellationToken)
-                    .ConfigureAwait(false);
+                await context.Database.ExecuteSqlRawAsync(sqlCreateOutputTableCopy, cancellationToken).ConfigureAwait(false);
             }
             else
             {
                 context.Database.ExecuteSqlRaw(sqlCreateOutputTableCopy);
             }
+
+            if (tableInfo.TimeStampColumnName != null)
+            {
+                var sqlAddColumn = SqlQueryBuilder.AddColumn(tableInfo.FullTempOutputTableName, tableInfo.TimeStampColumnName, TableInfo.TimeStampOutColumnType);
+                if (isAsync)
+                {
+                    await context.Database.ExecuteSqlRawAsync(sqlAddColumn, cancellationToken).ConfigureAwait(false);
+                }
+                else
+                {
+                    context.Database.ExecuteSqlRaw(sqlAddColumn);
+                }
+            }
+
+            if (operationType == OperationType.InsertOrUpdateOrDelete)
+            {
+                // Output returns all changes including Deleted rows with all NULL values, so if TempOutput.Id col not Nullable it breaks
+                var sqlAlterTableColumnsToNullable = SqlQueryBuilder.AlterTableColumnsToNullable(tableInfo.FullTempOutputTableName, tableInfo);
+                if (isAsync)
+                {
+                    await context.Database.ExecuteSqlRawAsync(sqlAlterTableColumnsToNullable, cancellationToken).ConfigureAwait(false);
+                }
+                else
+                {
+                    context.Database.ExecuteSqlRaw(sqlAlterTableColumnsToNullable);
+                }
+            }
         }
 
-        if (tableInfo.BulkConfig.CustomSourceTableName == null)
-        {
-            if (isAsync)
-            {
-                await InsertAsync(context, type, entities, tableInfo, progress, cancellationToken).ConfigureAwait(false);
-            }
-            else
-            {
-                Insert(context, type, entities, tableInfo, progress);
-            }
-        }
-
+        bool keepIdentity = tableInfo.BulkConfig.SqlBulkCopyOptions.HasFlag(SqlBulkCopyOptions.KeepIdentity);
         try
         {
-            var sqlMergeTable = SqlQueryBuilderMySql.MergeTable<T>(tableInfo, operationType);
+            if (tableInfo.BulkConfig.CustomSourceTableName == null)
+            {
+                if (isAsync)
+                {
+                    await InsertAsync(context, type, entities, tableInfo, progress, cancellationToken).ConfigureAwait(false);
+                }
+                else
+                {
+                    Insert(context, type, entities, tableInfo, progress);
+                }
+            }
+
+            if (keepIdentity && tableInfo.HasIdentity)
+            {
+                var sqlSetIdentityInsertTrue = SqlQueryBuilder.SetIdentityInsert(tableInfo.FullTableName, true);
+                if (isAsync)
+                {
+                    await context.Database.OpenConnectionAsync(cancellationToken).ConfigureAwait(false);
+                    await context.Database.ExecuteSqlRawAsync(sqlSetIdentityInsertTrue, cancellationToken).ConfigureAwait(false);
+                }
+                else
+                {
+                    context.Database.OpenConnection();
+                    context.Database.ExecuteSqlRaw(sqlSetIdentityInsertTrue);
+                }
+            }
+
+            var (sql, parameters) = SqlQueryBuilder.MergeTable<T>(context, tableInfo, operationType, entityPropertyWithDefaultValue);
             if (isAsync)
             {
-                await context.Database.ExecuteSqlRawAsync(sqlMergeTable, cancellationToken).ConfigureAwait(false);
+                await context.Database.ExecuteSqlRawAsync(sql, parameters, cancellationToken).ConfigureAwait(false);
             }
             else
             {
-                context.Database.ExecuteSqlRaw(sqlMergeTable);
+                context.Database.ExecuteSqlRaw(sql, parameters);
             }
+
             if (tableInfo.CreatedOutputTable)
             {
                 if (isAsync)
@@ -201,39 +262,14 @@ public class MySqLAdapter : ISqlOperationsAdapter
                     tableInfo.LoadOutputDataAsync(context, type, entities, tableInfo, isAsync: false, cancellationToken).GetAwaiter().GetResult();
                 }
             }
-            if (hasExistingTransaction == false && !tableInfo.BulkConfig.IncludeGraph)
-            {
-                if (isAsync)
-                {
-                    await transaction.CommitAsync(cancellationToken);
-                }
-                else
-                {
-                    transaction.Commit();
-                }
-            }
         }
         finally
         {
-            if (doDropUniqueConstrain)
-            {
-                string dropUniqueConstrain = SqlQueryBuilderMySql.DropUniqueConstrain(tableInfo);
-                if (isAsync)
-                {
-                    await context.Database.ExecuteSqlRawAsync(dropUniqueConstrain, cancellationToken)
-                        .ConfigureAwait(false);
-                }
-                else
-                {
-                    context.Database.ExecuteSqlRaw(dropUniqueConstrain);
-                }
-            }
-
             if (!tableInfo.BulkConfig.UseTempDB)
             {
                 if (tableInfo.CreatedOutputTable)
                 {
-                    var sqlDropOutputTable = SqlQueryBuilderMySql.DropTable(tableInfo.FullTempOutputTableName, tableInfo.InsertToTempTable);
+                    var sqlDropOutputTable = SqlQueryBuilder.DropTable(tableInfo.FullTempOutputTableName, tableInfo.BulkConfig.UseTempDB);
                     if (isAsync)
                     {
                         await context.Database.ExecuteSqlRawAsync(sqlDropOutputTable, cancellationToken).ConfigureAwait(false);
@@ -246,7 +282,7 @@ public class MySqLAdapter : ISqlOperationsAdapter
                 }
                 if (tableInfo.BulkConfig.CustomSourceTableName == null)
                 {
-                    var sqlDropTable = SqlQueryBuilderMySql.DropTable(tableInfo.FullTempTableName, tableInfo.InsertToTempTable);
+                    var sqlDropTable = SqlQueryBuilder.DropTable(tableInfo.FullTempTableName, tableInfo.BulkConfig.UseTempDB);
                     if (isAsync)
                     {
                         await context.Database.ExecuteSqlRawAsync(sqlDropTable, cancellationToken).ConfigureAwait(false);
@@ -256,102 +292,175 @@ public class MySqLAdapter : ISqlOperationsAdapter
                         context.Database.ExecuteSqlRaw(sqlDropTable);
                     }
                 }
-                if (hasExistingTransaction == false && !tableInfo.BulkConfig.IncludeGraph)
+            }
+
+            if (keepIdentity && tableInfo.HasIdentity)
+            {
+                var sqlSetIdentityInsertFalse = SqlQueryBuilder.SetIdentityInsert(tableInfo.FullTableName, false);
+                if (isAsync)
                 {
-                    if (isAsync)
-                    {
-                        await transaction.DisposeAsync();
-                    }
-                    else
-                    {
-                        transaction.Dispose();
-                    }
+                    await context.Database.ExecuteSqlRawAsync(sqlSetIdentityInsertFalse, cancellationToken).ConfigureAwait(false);
                 }
-            }    
+                else
+                {
+                    context.Database.ExecuteSqlRaw(sqlSetIdentityInsertFalse);
+                }
+                context.Database.CloseConnection();
+            }
         }
-        
-        
-    }
-    /// <inheritdoc/>
-    public void Read<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, Action<decimal>? progress) where T : class
-    {
-        throw new NotImplementedException();
     }
 
     /// <inheritdoc/>
-    public Task ReadAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, Action<decimal>? progress,
-        CancellationToken cancellationToken) where T : class
+    public void Read<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, Action<decimal>? progress) where T : class
     {
-        throw new NotImplementedException();
+        ReadAsync(context, type, entities, tableInfo, progress, isAsync: false, CancellationToken.None).GetAwaiter().GetResult();
+    }
+
+    /// <inheritdoc/>
+    public async Task ReadAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, Action<decimal>? progress, CancellationToken cancellationToken) where T : class
+    {
+        await ReadAsync(context, type, entities, tableInfo, progress, isAsync: true, cancellationToken).ConfigureAwait(false);
+    }
+
+    /// <inheritdoc/>
+    protected async Task ReadAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, Action<decimal>? progress, bool isAsync, CancellationToken cancellationToken) where T : class
+    {
+        Dictionary<string, string> previousPropertyColumnNamesDict = tableInfo.ConfigureBulkReadTableInfo();
+
+        var sqlCreateTableCopy = SqlQueryBuilder.CreateTableCopy(tableInfo.FullTableName, tableInfo.FullTempTableName, tableInfo);
+        if (isAsync)
+        {
+            await context.Database.ExecuteSqlRawAsync(sqlCreateTableCopy, cancellationToken).ConfigureAwait(false);
+        }
+        else
+        {
+            context.Database.ExecuteSqlRaw(sqlCreateTableCopy);
+        }
+
+        try
+        {
+            if (isAsync)
+            {
+                await InsertAsync(context, type, entities, tableInfo, progress, cancellationToken).ConfigureAwait(false);
+            }
+            else
+            {
+                InsertAsync(context, type, entities, tableInfo, progress, isAsync: false, cancellationToken).GetAwaiter().GetResult();
+            }
+
+            tableInfo.PropertyColumnNamesDict = tableInfo.OutputPropertyColumnNamesDict;
+
+            var sqlSelectJoinTable = SqlQueryBuilder.SelectJoinTable(tableInfo);
+
+            tableInfo.PropertyColumnNamesDict = previousPropertyColumnNamesDict; // TODO Consider refactor and integrate with TimeStampPropertyName, also check for Calculated props.
+                                                                                 // Output only PropertisToInclude and for getting Id with SetOutputIdentity
+            if (tableInfo.TimeStampPropertyName != null && !tableInfo.PropertyColumnNamesDict.ContainsKey(tableInfo.TimeStampPropertyName) && tableInfo.TimeStampColumnName is not null)
+            {
+                tableInfo.PropertyColumnNamesDict.Add(tableInfo.TimeStampPropertyName, tableInfo.TimeStampColumnName);
+            }
+
+            List<T> existingEntities = tableInfo.LoadOutputEntities<T>(context, type, sqlSelectJoinTable);
+            tableInfo.UpdateReadEntities(entities, existingEntities, context);
+
+            if (tableInfo.TimeStampPropertyName != null && !tableInfo.PropertyColumnNamesDict.ContainsKey(tableInfo.TimeStampPropertyName))
+            {
+                tableInfo.PropertyColumnNamesDict.Remove(tableInfo.TimeStampPropertyName);
+            }
+        }
+        finally
+        {
+            if (!tableInfo.BulkConfig.UseTempDB)
+            {
+                var sqlDropTable = SqlQueryBuilder.DropTable(tableInfo.FullTempTableName, tableInfo.BulkConfig.UseTempDB);
+                if (isAsync)
+                {
+                    await context.Database.ExecuteSqlRawAsync(sqlDropTable, cancellationToken).ConfigureAwait(false);
+                }
+                else
+                {
+                    context.Database.ExecuteSqlRaw(sqlDropTable);
+                }
+            }
+        }
     }
 
     /// <inheritdoc/>
     public void Truncate(DbContext context, TableInfo tableInfo)
     {
-        throw new NotImplementedException();
+        var sqlTruncateTable = SqlQueryBuilder.TruncateTable(tableInfo.FullTableName);
+        context.Database.ExecuteSqlRaw(sqlTruncateTable);
     }
 
     /// <inheritdoc/>
-    public Task TruncateAsync(DbContext context, TableInfo tableInfo, CancellationToken cancellationToken)
+    public async Task TruncateAsync(DbContext context, TableInfo tableInfo, CancellationToken cancellationToken)
     {
-        throw new NotImplementedException();
+        var sqlTruncateTable = SqlQueryBuilder.TruncateTable(tableInfo.FullTableName);
+        await context.Database.ExecuteSqlRawAsync(sqlTruncateTable, cancellationToken).ConfigureAwait(false);
     }
     #endregion
-    #region Connection
 
-    internal static async Task<(MySqlConnection, bool)> OpenAndGetMySqlConnectionAsync(DbContext context, CancellationToken cancellationToken)
+    #region Connection
+    private static SqlBulkCopy GetSqlBulkCopy(SqlConnection sqlConnection, IDbContextTransaction? transaction, BulkConfig config)
     {
-        bool closeConnectionInternally = false;
-        var mySqlConnection = (MySqlConnection)context.Database.GetDbConnection();
-        if (mySqlConnection.State != ConnectionState.Open)
+        var sqlTransaction = transaction == null ? null : (SqlTransaction)transaction.GetUnderlyingTransaction(config);
+        var sqlBulkCopy = new SqlBulkCopy(sqlConnection, config.SqlBulkCopyOptions, sqlTransaction);
+        if (config.SqlBulkCopyColumnOrderHints != null)
         {
-            await mySqlConnection.OpenAsync(cancellationToken).ConfigureAwait(false);
-            closeConnectionInternally = true;
+            foreach(var hint in config.SqlBulkCopyColumnOrderHints)
+                sqlBulkCopy.ColumnOrderHints.Add(hint);
         }
-        return (mySqlConnection, closeConnectionInternally);
+        return sqlBulkCopy;
+    }
+    /// <summary>
+    /// Supports <see cref="Microsoft.Data.SqlClient.SqlBulkCopy"/>
+    /// </summary>
+    /// <typeparam name="T"></typeparam>
+    /// <param name="sqlBulkCopy"></param>
+    /// <param name="tableInfo"></param>
+    /// <param name="entities"></param>
+    /// <param name="setColumnMapping"></param>
+    /// <param name="progress"></param>
+    private static void SetSqlBulkCopyConfig<T>(Microsoft.Data.SqlClient.SqlBulkCopy sqlBulkCopy, TableInfo tableInfo, IList<T> entities, bool setColumnMapping, Action<decimal>? progress)
+    {
+        sqlBulkCopy.DestinationTableName = tableInfo.InsertToTempTable ? tableInfo.FullTempTableName : tableInfo.FullTableName;
+        sqlBulkCopy.BatchSize = tableInfo.BulkConfig.BatchSize;
+        sqlBulkCopy.NotifyAfter = tableInfo.BulkConfig.NotifyAfter ?? tableInfo.BulkConfig.BatchSize;
+        sqlBulkCopy.SqlRowsCopied += (sender, e) =>
+        {
+            progress?.Invoke(ProgressHelper.GetProgress(entities.Count, e.RowsCopied)); // round to 4 decimal places
+        };
+        sqlBulkCopy.BulkCopyTimeout = tableInfo.BulkConfig.BulkCopyTimeout ?? sqlBulkCopy.BulkCopyTimeout;
+        sqlBulkCopy.EnableStreaming = tableInfo.BulkConfig.EnableStreaming;
+
+        if (setColumnMapping)
+        {
+            foreach (var element in tableInfo.PropertyColumnNamesDict)
+            {
+                sqlBulkCopy.ColumnMappings.Add(element.Key, element.Value);
+            }
+        }
     }
 
-    internal static (MySqlConnection, bool) OpenAndGetMySqlConnection(DbContext context)
-    {
-        bool closeConnectionInternally = false;
-        var myConnection = (MySqlConnection)context.Database.GetDbConnection();
-        if (myConnection.State != ConnectionState.Open)
-        {
-            myConnection.Open();
-            closeConnectionInternally = true;
-        }
-        return (myConnection, closeConnectionInternally);
-    }
     #endregion
-    #region Connection
-    private static MySqlBulkCopy GetMySqlBulkCopy(MySqlConnection mySqlConnection, IDbContextTransaction? transaction, BulkConfig config)
-    {
-        var mySqlTransaction = transaction == null ? null : (MySqlTransaction)transaction.GetUnderlyingTransaction(config);
-        var mySqlBulkCopy = new MySqlBulkCopy(mySqlConnection, mySqlTransaction);
-        
-        return mySqlBulkCopy;
-    }
-    #endregion
+
     #region DataTable
     /// <summary>
-    /// Supports <see cref="MySqlBulkCopy"/>
+    /// Supports <see cref="SqlBulkCopy"/>
     /// </summary>
     /// <typeparam name="T"></typeparam>
     /// <param name="context"></param>
     /// <param name="type"></param>
     /// <param name="entities"></param>
-    /// <param name="mySqlBulkCopy"></param>
+    /// <param name="sqlBulkCopy"></param>
     /// <param name="tableInfo"></param>
     /// <returns></returns>
-    public static DataTable GetDataTable<T>(DbContext context, Type type, IList<T> entities, MySqlBulkCopy mySqlBulkCopy, TableInfo tableInfo)
+    public static DataTable GetDataTable<T>(DbContext context, Type type, IList<T> entities, SqlBulkCopy sqlBulkCopy, TableInfo tableInfo)
     {
         DataTable dataTable = InnerGetDataTable(context, ref type, entities, tableInfo);
 
-        int sourceOrdinal = 0;
         foreach (DataColumn item in dataTable.Columns)  //Add mapping
         {
-            mySqlBulkCopy.ColumnMappings.Add(new MySqlBulkCopyColumnMapping(sourceOrdinal,item.ColumnName));
-            sourceOrdinal++;
+            sqlBulkCopy.ColumnMappings.Add(item.ColumnName, item.ColumnName);
         }
         return dataTable;
     }
@@ -371,9 +480,10 @@ public class MySqLAdapter : ISqlOperationsAdapter
         var columnsDict = new Dictionary<string, object?>();
         var ownedEntitiesMappedProperties = new HashSet<string>();
 
-        var databaseType = SqlAdaptersMapping.GetDatabaseType(context);
-        var isMySql = databaseType == DbServer.MySQL;
-        
+        var databaseType = SqlAdaptersMapping.GetDatabaseType();
+        var isSqlServer = databaseType == DbServer.SQLServer;
+        var sqlServerBytesWriter = new SqlServerBytesWriter();
+
         var objectIdentifier = tableInfo.ObjectIdentifier;
         type = tableInfo.HasAbstractList ? entities[0]!.GetType() : type;
         var entityType = context.Model.FindEntityType(type) ?? throw new ArgumentException($"Unable to determine entity type from given type - {type.Name}");
@@ -415,7 +525,7 @@ public class MySqLAdapter : ISqlOperationsAdapter
                     propertyType = underlyingType;
                 }
 
-                if (isMySql && (propertyType == typeof(Geometry) || propertyType.IsSubclassOf(typeof(Geometry))))
+                if (isSqlServer && (propertyType == typeof(Geometry) || propertyType.IsSubclassOf(typeof(Geometry))))
                 {
                     propertyType = typeof(byte[]);
                     tableInfo.HasSpatialType = true;
@@ -424,7 +534,7 @@ public class MySqLAdapter : ISqlOperationsAdapter
                         throw new InvalidOperationException("OnCompare properties Config can not be set for Entity with Spatial types like 'Geometry'");
                     }
                 }
-                if (isMySql && (propertyType == typeof(HierarchyId) || propertyType.IsSubclassOf(typeof(HierarchyId))))
+                if (isSqlServer && (propertyType == typeof(HierarchyId) || propertyType.IsSubclassOf(typeof(HierarchyId))))
                 {
                     propertyType = typeof(byte[]);
                 }
@@ -454,12 +564,12 @@ public class MySqLAdapter : ISqlOperationsAdapter
                     propertyType = underlyingType;
                 }
 
-                if (propertyType == typeof(Geometry) && isMySql)
+                if (propertyType == typeof(Geometry) && isSqlServer)
                 {
                     propertyType = typeof(byte[]);
                 }
 
-                if (propertyType == typeof(HierarchyId) && isMySql)
+                if (propertyType == typeof(HierarchyId) && isSqlServer)
                 {
                     propertyType = typeof(byte[]);
                 }
@@ -510,14 +620,31 @@ public class MySqLAdapter : ISqlOperationsAdapter
                             {
                                 var convertor = tableInfo.ConvertibleColumnConverterDict[propertyName];
                                 var underlyingType = Nullable.GetUnderlyingType(convertor.ProviderClrType) ?? convertor.ProviderClrType;
+          
                                 dataTable.Columns.Add(columnName, underlyingType);
                             }
                             else
                             {
                                 var ownedPropertyType = Nullable.GetUnderlyingType(innerProperty.PropertyType) ?? innerProperty.PropertyType;
+                                
+                                if (isSqlServer && (ownedPropertyType == typeof(Geometry) || ownedPropertyType.IsSubclassOf(typeof(Geometry))))
+                                {
+                                    ownedPropertyType = typeof(byte[]);
+                                    tableInfo.HasSpatialType = true;
+                                    if (tableInfo.BulkConfig.PropertiesToIncludeOnCompare != null || tableInfo.BulkConfig.PropertiesToIncludeOnCompare != null)
+                                    {
+                                        throw new InvalidOperationException("OnCompare properties Config can not be set for Entity with Spatial types like 'Geometry'");
+                                    }
+                                }
+
+                                if (isSqlServer && (ownedPropertyType == typeof(HierarchyId) || ownedPropertyType.IsSubclassOf(typeof(HierarchyId))))
+                                {
+                                    ownedPropertyType = typeof(byte[]);
+                                }
+
                                 dataTable.Columns.Add(columnName, ownedPropertyType);
                             }
-
+  ;
                             columnsDict.Add(property.Name + "_" + innerProperty.Name, null);
                         }
                     }
@@ -547,12 +674,12 @@ public class MySqLAdapter : ISqlOperationsAdapter
                     propertyType = underlyingType;
                 }
 
-                if (isMySql && (propertyType == typeof(Geometry) || propertyType.IsSubclassOf(typeof(Geometry))))
+                if (isSqlServer && (propertyType == typeof(Geometry) || propertyType.IsSubclassOf(typeof(Geometry))))
                 {
                     propertyType = typeof(byte[]);
                 }
 
-                if (isMySql && (propertyType == typeof(HierarchyId) || propertyType.IsSubclassOf(typeof(HierarchyId))))
+                if (isSqlServer && (propertyType == typeof(HierarchyId) || propertyType.IsSubclassOf(typeof(HierarchyId))))
                 {
                     propertyType = typeof(byte[]);
                 }
@@ -589,7 +716,7 @@ public class MySqLAdapter : ISqlOperationsAdapter
                     && tableInfo.DefaultValueProperties.Contains(property.Name);
 
                 if (tableInfo.BulkConfig.DateTime2PrecisionForceRound
-                    && isMySql
+                    && isSqlServer
                     && tableInfo.DateTime2PropertiesPrecisionLessThen7Dict.ContainsKey(property.Name))
                 {
                     DateTime? dateTimePropertyValue = (DateTime?)propertyValue;
@@ -614,9 +741,20 @@ public class MySqLAdapter : ISqlOperationsAdapter
                     string columnName = tableInfo.ConvertiblePropertyColumnDict[property.Name];
                     propertyValue = tableInfo.ConvertibleColumnConverterDict[columnName].ConvertToProvider.Invoke(propertyValue);
                 }
-                //TODO: Hamdling special types
 
-                if (propertyValue is HierarchyId hierarchyValue && isMySql)
+                if (tableInfo.HasSpatialType && propertyValue is Geometry geometryValue)
+                {
+                    geometryValue.SRID = tableInfo.BulkConfig.SRID;
+
+                    if (tableInfo.PropertyColumnNamesDict.ContainsKey(property.Name))
+                    {
+                        sqlServerBytesWriter.IsGeography = tableInfo.ColumnNamesTypesDict[tableInfo.PropertyColumnNamesDict[property.Name]] == "geography"; // "geography" type is default, otherwise it's "geometry" type
+                    }
+
+                    propertyValue = sqlServerBytesWriter.Write(geometryValue);
+                }
+
+                if (propertyValue is HierarchyId hierarchyValue && isSqlServer)
                 {
                     using MemoryStream memStream = new();
                     using BinaryWriter binWriter = new(memStream);
@@ -654,6 +792,17 @@ public class MySqLAdapter : ISqlOperationsAdapter
                         {
                             var converter = tableInfo.ConvertibleColumnConverterDict[columnName];
                             columnsDict[columnName] = ownedPropertyValue == null ? null : converter.ConvertToProvider.Invoke(ownedPropertyValue);
+                        }
+                        else if(tableInfo.HasSpatialType && ownedPropertyValue is Geometry ownedGeometryValue)
+                        {
+                            ownedGeometryValue.SRID = tableInfo.BulkConfig.SRID;
+
+                            if (tableInfo.PropertyColumnNamesDict.ContainsKey(property.Name))
+                            {
+                                sqlServerBytesWriter.IsGeography = tableInfo.ColumnNamesTypesDict[tableInfo.PropertyColumnNamesDict[property.Name]] == "geography"; // "geography" type is default, otherwise it's "geometry" type
+                            }
+
+                            columnsDict[columnName] = sqlServerBytesWriter.Write(ownedGeometryValue);
                         }
                         else
                         {
