+++ EFCore.BulkExtensions.SqlServer/SqlServerDialect.cs
@@ -1,0 +1,13 @@
+﻿using EFCore.BulkExtensions.SqlAdapters;
+using Microsoft.EntityFrameworkCore;
+using System;
+using System.Collections.Generic;
+using System.Data;
+using System.Linq.Expressions;
+
+namespace EFCore.BulkExtensions.SqlAdapters.SQLServer;
+
+/// <inheritdoc/>
+public class SqlServerDialect : SqlDefaultDialect
+{
+}
