+++ EFCore.BulkExtensions.Tests/SqlQueryBuilderPostgreSqlTests.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using EFCore.BulkExtensions.SQLAdapters.PostgreSql;
+using EFCore.BulkExtensions.SqlAdapters.PostgreSql;
 using Xunit;
 
 namespace EFCore.BulkExtensions.Tests;
