+++ EFCore.BulkExtensions/DbContextBulkTransactionGraphUtil.cs
@@ -31,7 +31,7 @@ internal static class DbContextBulkTransactionGraphUtil
             throw new InvalidBulkConfigException($"{nameof(BulkConfig)}.{nameof(BulkConfig.IncludeGraph)} only supports Insert or Update operations.");
 
         // Sqlite bulk merge adapter does not support multiple objects of the same type with a zero value primary key
-        if (SqlAdaptersMapping.GetDatabaseType(context) == DbServer.SQLite)
+        if (SqlAdaptersMapping.GetDatabaseType() == DbServer.SQLite)
             throw new NotSupportedException("Sqlite is not currently supported due to its BulkInsert implementation.");
 
         bulkConfig.PreserveInsertOrder = true; // Required for SetOutputIdentity ('true' is default but here explicitly assigned again in case it was changed to 'false' in BulkConfing)
