+++ EFCore.BulkExtensions/DbContextBulkTransactionSaveChanges.cs
@@ -1,5 +1,4 @@
-﻿using EFCore.BulkExtensions.SQLAdapters.SQLServer;
-using Microsoft.Data.SqlClient;
+﻿using Microsoft.Data.SqlClient;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.ChangeTracking;
 using System;
