+++ EFCore.BulkExtensions/EFCore.BulkExtensions.csproj
@@ -20,7 +20,7 @@
     <AssemblyVersion>6.5.6.0</AssemblyVersion>
     <FileVersion>6.5.6.0</FileVersion>
     <SignAssembly>true</SignAssembly>
-    <AssemblyOriginatorKeyFile>Keys\EFCore.BulkExtensions.snk</AssemblyOriginatorKeyFile>
+    <AssemblyOriginatorKeyFile>..\Keys\EFCore.BulkExtensions.snk</AssemblyOriginatorKeyFile>
     <DelaySign>false</DelaySign>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <LangVersion>latest</LangVersion>
@@ -28,13 +28,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="EntityFrameworkCore.SqlServer.HierarchyId" Version="3.0.1" />
     <PackageReference Include="Microsoft.Data.SqlClient" Version="2.1.4" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="6.0.8" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite.Core" Version="6.0.8" />
-    <PackageReference Include="NetTopologySuite.IO.SqlServerBytes" Version="2.1.0" />
-    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="6.0.6" />
-    <PackageReference Include="Pomelo.EntityFrameworkCore.MySql" Version="6.0.2" />
     <PackageReference Include="System.Diagnostics.DiagnosticSource" Version="6.0.0" />
   </ItemGroup>
 
@@ -45,4 +40,27 @@
     </None>
   </ItemGroup>
 
+    <ItemGroup>
+        <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
+            <_Parameter1>
+                EFCore.BulkExtensions.MySql, PublicKey=0024000004800000940000000602000000240000525341310004000001000100a1b05d950ba6c0ef9dbea919646d1bf24a906258703a2790dce9e570786825196caace78adc77c50a95b9222f5d38192453fb080f29af0ca297f29b40996c5df2f44226e91705a799ed853590e49b06ed66a34bc4552e78609e45f303d8f77c5260a990bca188e1dee17d89c7f92e2f1f62b9addcb59857e88f86b7e375a6fb4
+            </_Parameter1>
+        </AssemblyAttribute>
+        <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
+            <_Parameter1>
+                EFCore.BulkExtensions.PostgreSql, PublicKey=0024000004800000940000000602000000240000525341310004000001000100a1b05d950ba6c0ef9dbea919646d1bf24a906258703a2790dce9e570786825196caace78adc77c50a95b9222f5d38192453fb080f29af0ca297f29b40996c5df2f44226e91705a799ed853590e49b06ed66a34bc4552e78609e45f303d8f77c5260a990bca188e1dee17d89c7f92e2f1f62b9addcb59857e88f86b7e375a6fb4
+            </_Parameter1>
+        </AssemblyAttribute>
+        <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
+            <_Parameter1>
+                EFCore.BulkExtensions.SQLite, PublicKey=0024000004800000940000000602000000240000525341310004000001000100a1b05d950ba6c0ef9dbea919646d1bf24a906258703a2790dce9e570786825196caace78adc77c50a95b9222f5d38192453fb080f29af0ca297f29b40996c5df2f44226e91705a799ed853590e49b06ed66a34bc4552e78609e45f303d8f77c5260a990bca188e1dee17d89c7f92e2f1f62b9addcb59857e88f86b7e375a6fb4
+            </_Parameter1>
+        </AssemblyAttribute>
+        <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
+            <_Parameter1>
+                EFCore.BulkExtensions.SqlServer, PublicKey=0024000004800000940000000602000000240000525341310004000001000100a1b05d950ba6c0ef9dbea919646d1bf24a906258703a2790dce9e570786825196caace78adc77c50a95b9222f5d38192453fb080f29af0ca297f29b40996c5df2f44226e91705a799ed853590e49b06ed66a34bc4552e78609e45f303d8f77c5260a990bca188e1dee17d89c7f92e2f1f62b9addcb59857e88f86b7e375a6fb4
+            </_Parameter1>
+        </AssemblyAttribute>
+    </ItemGroup>
+    
 </Project>
