--- EFCore.BulkExtensions/SQLAdapters/MySql/MySqlDialect.cs
@@ -1,9 +1,0 @@
-﻿using EFCore.BulkExtensions.SQLAdapters.SQLServer;
-
-namespace EFCore.BulkExtensions.SqlAdapters.MySql;
-
-/// <inheritdoc/>
-public class MySqlDialect : SqlServerDialect
-{
-    
-}
