--- EFCore.BulkExtensions/SQLAdapters/PostgreSql/PostgreSqlDialect.cs
@@ -1,8 +1,0 @@
-﻿using EFCore.BulkExtensions.SQLAdapters.SQLServer;
-
-namespace EFCore.BulkExtensions.SQLAdapters.PostgreSql;
-
-/// <inheritdoc/>
-public class PostgreSqlDialect : SqlServerDialect
-{
-}
