+++ EFCore.BulkExtensions/SqlBulkOperation.cs
@@ -52,25 +52,25 @@ internal static class SqlBulkOperation
 
     public static void Insert<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, Action<decimal>? progress)
     {
-        var adapter = SqlAdaptersMapping.CreateBulkOperationsAdapter(context);
+        var adapter = SqlAdaptersMapping.CreateBulkOperationsAdapter();
         adapter.Insert(context, type, entities, tableInfo, progress);
     }
 
     public static async Task InsertAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, Action<decimal>? progress, CancellationToken cancellationToken)
     {
-        var adapter = SqlAdaptersMapping.CreateBulkOperationsAdapter(context);
+        var adapter = SqlAdaptersMapping.CreateBulkOperationsAdapter();
         await adapter.InsertAsync(context, type, entities, tableInfo, progress, cancellationToken);
     }
 
     public static void Merge<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, OperationType operationType, Action<decimal>? progress) where T : class
     {
-        var adapter = SqlAdaptersMapping.CreateBulkOperationsAdapter(context);
+        var adapter = SqlAdaptersMapping.CreateBulkOperationsAdapter();
         adapter.Merge(context, type, entities, tableInfo, operationType, progress);
     }
 
     public static async Task MergeAsync<T>(DbContext context, Type type, IList<T> entities, TableInfo tableInfo, OperationType operationType, Action<decimal>? progress, CancellationToken cancellationToken = default) where T : class
     {
-        var adapter = SqlAdaptersMapping.CreateBulkOperationsAdapter(context);
+        var adapter = SqlAdaptersMapping.CreateBulkOperationsAdapter();
         await adapter.MergeAsync(context, type, entities, tableInfo, operationType, progress, cancellationToken);
     }
 
@@ -80,7 +80,7 @@ internal static class SqlBulkOperation
         {
             context.Database.ExecuteSqlRaw(SqlQueryBuilder.DropTable(tableInfo.FullTempTableName, tableInfo.BulkConfig.UseTempDB));
         }
-        var adapter = SqlAdaptersMapping.CreateBulkOperationsAdapter(context);
+        var adapter = SqlAdaptersMapping.CreateBulkOperationsAdapter();
         adapter.Read(context, type, entities, tableInfo, progress);
     }
 
@@ -90,19 +90,19 @@ internal static class SqlBulkOperation
         {
             await context.Database.ExecuteSqlRawAsync(SqlQueryBuilder.DropTable(tableInfo.FullTempTableName, tableInfo.BulkConfig.UseTempDB), cancellationToken).ConfigureAwait(false);
         }
-        var adapter = SqlAdaptersMapping.CreateBulkOperationsAdapter(context);
+        var adapter = SqlAdaptersMapping.CreateBulkOperationsAdapter();
         await adapter.ReadAsync(context, type, entities, tableInfo, progress, cancellationToken);
     }
 
     public static void Truncate(DbContext context, TableInfo tableInfo)
     {
-        var adapter = SqlAdaptersMapping.CreateBulkOperationsAdapter(context);
+        var adapter = SqlAdaptersMapping.CreateBulkOperationsAdapter();
         adapter.Truncate(context, tableInfo);
     }
 
     public static async Task TruncateAsync(DbContext context, TableInfo tableInfo, CancellationToken cancellationToken)
     {
-        var adapter = SqlAdaptersMapping.CreateBulkOperationsAdapter(context);
+        var adapter = SqlAdaptersMapping.CreateBulkOperationsAdapter();
         await adapter.TruncateAsync(context, tableInfo, cancellationToken).ConfigureAwait(false);
     }
 }
